Support for multiple mailbox protocols.
In this solution I accept that a mailbox read request (e.g. FP-RD) for
a given mailbox protocol can return data from any other mailbox
protocol running at the same time. The data returned by a read datagram
is therefore stored in a separate buffer for each mailbox protocol
instead of the datagram data buffer. The mailbox state machines will
check and fetch the data from their own buffer instead of the datagram
buffer (that is no longer used for mailbox read data). A check_mbox
flag is introduced to track when a given slave has an ongoing mailbox
read request. In normal case the mailbox state machine will run as
previously if no mailbox read request is ongoing, but if a mailbox
read-request is ongoing (check_mbox flag is set) it will check its own
mailbox buffer (as the ongoing mailbox read request might have returned
its data) and otherwise wait until the read request is done and it gets
the opportunity to reserve the mailbox for its own read request.
* * *
The multiple mailbox support patch introduced a read_mbox_busy flag that
is not cleared in a special case where an ecrt_master_deactivate is called
during a PDO download to a slave with PDO configuration enabled.